backend/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.1.3

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.41

# File handling
cloudinary==1.37.0
Pillow==10.2.0

# Payment integration
requests==2.31.0

# Validation
marshmallow==3.21.1
email-validator==2.1.1

# Security
bcrypt==4.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==24.3.0
flake8==7.0.0
